+++ nion/ui/UserInterface.py
@@ -2903,6 +2903,13 @@ class ToleranceType(enum.IntEnum):
     CURSOR = 1
 
 
+class TruncateModeType(enum.IntEnum):
+    LEFT = 0
+    RIGHT = 1
+    MIDDLE = 2
+    NONE = 3
+
+
 class UserInterface(abc.ABC):
 
     @abc.abstractmethod
@@ -3085,6 +3092,10 @@ class UserInterface(abc.ABC):
 
     @abc.abstractmethod
     def get_font_metrics(self, font: str, text: str) -> FontMetrics:
+        ...
+
+    @abc.abstractmethod
+    def truncate_string_to_width(self, font_str: str, text: str, pixel_width: int, mode: TruncateModeType) -> str:
         ...
 
     @abc.abstractmethod